+++ scripts/download-sdks.ps1
@@ -34,10 +34,17 @@ if (-not (Test-Path $outDir))
     New-Item $outDir -ItemType Directory > $null
 }
 
-$sdks = @("Android", "IOS", "Linux", "Mac", "Win64")
+$sdks = @("Android", "IOS", "Linux", "Mac", "Win64-Crashpad", "Win64-Breakpad")
 foreach ($sdk in $sdks)
 {
     $sdkDir = "$outDir/$sdk"
+
+    if ($sdk.StartsWith('Win64'))
+    {
+        $winSdk, $crashBackend = $sdk.Split("-")
+        $sdkDir = "$outDir/$winSdk/$crashBackend"
+    }  
+
     Write-Host "Downloading $sdk SDK to $sdkDir ..."
     if (Test-Path $sdkDir)
     {